Initial commit
A working example using Qt and UDP multicast to share screen.

Server constructs a multicast ip with 224.0.1.x while x is the least
signicant part of its ip. Server broadcasts this ip at 224.0.0.1 and
multicasts its screen to 224.0.1.x.

Client listens to 224.0.0.1 for server addresses. After a server is
selected, client joins to 224.0.1.x and displays the received images